requirements.txt
pyTelegramBotAPI
paho-mqtt
